Added a vertical nav component. Moved route logic
